Add StmtM applicative for statement evaluation
This adds the Eval.Statement module with the StmtM applicative
functor, in which evaluation of top-level statements takes place.
It is not a monad, because it is based on the Validation type --
the idea is that a statement cannot depend on the result of a
previous statement anyway, so we may as well collect all the
errors, and that requires something like Validation, which cannot
be a Monad. I think having top-level statements be independent in
this way is actually a nice limitation of the language.

This also adds the Control.Applicative.Writer module, with a
version of the WriterT monad transformer adjusted to relax its
Monad constraints to Applicative (and even to Functor) wherever
possible.