Add ValidationT, and rewrite StmtT monad to use it
Instead of depending on the validation package, we now define our
own ValidationT applicative transformer. (For convenience, we also
define a Monad instance -- this is frowned upon, but, again,
extremely useful.) This allowed me to simplify the implementation
of StmtT somewhat. In particular, I no longer need the hand-written
Applicative and Monad instances.