Remove List signature, make List = Vector instead
This commit removes the List.hsig module signature, and replaces it
with its Vector implementation. From now on, List is just a type
synonym for Vector -- no Backpack involved.

This is simpler, and it means cabal is a lot quicker to rebuild the
project (fewer opportunities for spurious reconfigures and
no-op rebuilds). I wasn't really making use of the signature
abstraction anyway.