Enable optimization again
Now that I have an executable to run, why not have it run fast?
I had optimization off before to save time rebuilding, but now I am
switching to using ghcid for checking that it compiles, so the
speed of compiling the project does not matter as much. (It is slow
though! Especially when cabal thinks a reconfigure is needed.)