Parse applications more liberally, then validate
This commit changes the parsing of applications so that any name
followed by some arguments is parsed, and then we check whether the
name is a function and how many arguments it takes -- as opposed to
checking whether the name refers to a function before parsing the
arguments. This allows us to give better error messages. (They're
not that much better now, but I'll improve them.)