Export 'toShakeOptions', and don't 'want' targets in 'rules'
The 'toShakeOptions' function, previously internal to
"DependencyMonad", is now exported, for use by the test suite.

Additionally, 'rules' no longer calls 'want' for the targets given
in 'Options{optTargets}' ('run' still does, of course). This is so
that the test suite can re-use the 'Rules' for different sets of
targets.