Update the golden tests to work with the new API
The golden test suite now runs the shake rules using
'shakeWithDatabase' et al. This is more complicated than the
alternative of just calling 'run' a bunch of times, but more fun,
and possibly faster.

The tests will not pass yet, because there are YAML files lying
around that the new test harness does not know about.