Overhaul how targets are specified
Prior to this, in order to tell Thales what to build, you had to
specify, either on the command line or programmatically in the
'Options' struct, the *targets*, that is, the files that should be
*produced*. I find it more intuitive to specify, instead, the
template files -- i.e. the source files -- that should be built. In
general, I would rather pass files that exists, rather than files
that are supposed to be created, on the command line, especially
when the targets are not predefined, but rather are in principle
determined from the set of source files anyway.

With this patch, you must now specify some number of filepaths or
glob patterns that specify the template files to be built. You must
also specify the extension to be given to output files. From this
information, the build system figures out what the outputs should
be.

This way of doing things -- determining outputs from inputs -- is
not quite what Shake is designed around. The way I've hacked it is
to construct the reverse mapping, from outputs to inputs (which is
what Shake wants), outside of the rule definitions. See the source
(I should really add some comments ...)