Add --output-directory option
The `--output-directory` option specifies the directory where the
build products should be put. It defaults to the current directory
(which was effectively the previous behaviour). It is prepended to
the path for each build product (i.e. it does not replace the
directory compoment of the path).

The golden test harness now uses this to put the outputs from the
test run in a temporary directory, so that they do not clutter up
the test directory.