Re-organize golden files for new test harness
The golden test harness used to be able to automatically pick up a
yaml file with the same filename as a template file and bring the
names defined in the yaml file into scope when executing the
template. This is no longer possible, because the test suite uses
a different API which offers no way to introduce additional
top-level bindings. So each template has to import the yaml they
want explicitly.

The test harness also used to be able to match up a template file
with a directory of output files and corresponding yaml files, and
execute the template once for each, but this is no longer possible,
because of the above, so I had to rename and duplicate some files.