Use variable instead of literal for "body" field special behaviour
The 'body' field of a record that is the result of loading a file
that has a "document body", so to speak, namely a template or
markdown file -- e.g. '(load-template "blah.template" {}).body' --
is the field by which that "document body" is accessible. The name
of this field is hardcoded -- it might be nice to make it
configurable, but for now the least I can do it is put in in a
named variable instead of writing the literal "body" everywhere.