Parametrize 'ThingToBuild' over a functor, and add a few fields
This is essentially the Higher-Kinded Data pattern again. The idea
is that when we are parsing a 'ThingToBuild', i.e. parameters for
building a set of templates, from a configuration file (not
implemented yet) or from the command line, some fields are
optional ('ThingToBuild Maybe'), but these get filled in later from
the global 'Options' ('ThingToBuild Identity').