Add Display instance for Problem et al
Thales can now pretty-print errors encountered when evaluating a
template. This is accomplished in the Display instances for Problem
and ProblemDescription. There is also a new accompanying instance
for Value, as well as for the various auxiliary types such as
ArgumentErrors.

I have also split the definition of built-in functions into a new
module Eval.KnownFunction.

Additionally I have factored out some common code for
pretty-printing of lists and records.

There is not yet support for ANSI terminal markup, but I have that
planned. (The Display instance for ProblemWhere is kind of useless
without it -- in fact that was the reason I created ProblemWhere in
the first place!)