Make Record its own type
Previously a 'Record' in the template language could be represented
as either of two constructors of the 'Value' datatype, namely
'Record' for concrete records, and 'ExternalRecord' for records
whose fields are lazily loaded from some file. These now correspond
to the 'Concrete' and 'External' constructors of the new 'Record'
datatype. Now every type in the template language is represented by
exactly one data constructor of 'Value'. This makes some changes I
have in mind for checking of function arguments easier.