Completely rewrite checking of function arguments
Checking of the types of function arguments is now accomplished in
a much simpler way, with a much smaller and simultaneously more
expressive API. The idea is to *parse* a lists of dynamically-typed
values, where the result of a successful parse is the application
of the function. The basic combinator is 'argument', which means
'get the next argument, requiring it to be of this type'.

Aside from the simpler and less weird API, the big advantage of
this rewrite -- and my original motivation for it -- was that the
previous version's error messages were dodgy. I realized that I
needed to distinguish between errors that appeared after the parser
had committed to a particular alternative, and errors that appeared
immediately -- this distinction being one that every parser
cominbator library that cares about error messages has to pay
attention to.