Make 'Bindings' a synonym for 'HashMap etc.' again
I realized there was no point in having 'Bindings' be a newtype,
because all of the functions on it are exactly the same as HashMap,
modulo the 'Name' key newtype, which is easily worked around, and I
wasn't even using the 'Bindings' vs 'HashMap Text Value' consistently.
I kept the module, though, so as not to have to change so much
client code. Might remove it altogether someday.